+++ plone/app/layout/globals/layout.py
@@ -1,6 +1,7 @@
 import json
 
 from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.interfaces.controlpanel import ILinkSchema
 from Products.CMFPlone.interfaces.controlpanel import ISiteSchema
 from Products.Five.browser.metaconfigure import ViewMixinForTemplates
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
@@ -164,11 +165,8 @@ class LayoutPolicy(BrowserView):
             # skip first section since we already have that...
             if len(contentPath) > 1:
                 registry = getUtility(IRegistry)
-                try:
-                    depth = registry[
-                        'plone.app.layout.globals.bodyClass.depth']
-                except KeyError:
-                    depth = 4
+                depth = registry.get(
+                    'plone.app.layout.globals.bodyClass.depth', 4)
                 if depth > 1:
                     classes = ['subsection-%s' % contentPath[1]]
                     for section in contentPath[2:depth]:
@@ -181,7 +179,8 @@ class LayoutPolicy(BrowserView):
         else:
             body_classes.append('icons-off')
 
-        # permissions required. Useful to theme frontend and backend differently
+        # permissions required. Useful to theme frontend and backend
+        # differently
         permissions = []
         if not getattr(view, '__ac_permissions__', tuple()):
             permissions = ['none']
@@ -235,11 +234,13 @@ class LayoutPolicy(BrowserView):
                 pass
 
         # class for markspeciallinks pattern
-        properties = getToolByName(context, "portal_properties")
-        props = getattr(properties, 'site_properties')
+        registry = getUtility(IRegistry)
+        settings = registry.forInterface(ILinkSchema,
+                                         prefix="plone",
+                                         check=False)
 
-        msl = props.getProperty('mark_special_links', 'false')
-        elonw = props.getProperty('external_links_open_new_window', 'false')
+        msl = settings.mark_special_links
+        elonw = settings.external_links_open_new_window
         if msl == 'true' or elonw == 'true':
             body_classes.append('pat-markspeciallinks')